s5e8825-common: props: Disable vsync for CPU rendered Apps
* Choreographer VSync has some performance impact for CPU rendered apps
as the app has to wait for VSync to start a new frame draw. When the
frame draw time exceeds one VSync, this has a direct impact on the FPS.
debug.cpurend.vsync property controls the VSync for CPU rendered apps.

Disable VSync for CPU rendered app, by setting this property to false.

Change-Id: I8f30937e680078384a7a322dd8ce1cf8be373ebb
Signed-off-by: Ksawlii <ksawery.blaszczak@proton.me>