s5e8825-common: sepolicy: Declare hermesd as hal_server_domain
This is a very important fix that, if not applied, will cause hermesd to
report:

[   53.938621] !@ hermesd.cpp(109) Failed to register Weaver as a service (status: 2147483648)

And it won't work properly as a result.

It won't print a denial when doing this, and I couldn't spot anything in
the log indicating what the issue was, but simply switching to Permissive
would allow it start.

The fix was found by referencing the following commit:
https://github.com/GrapheneOS/device_google_gs201-sepolicy/commit/9cb1f625ba9a6b76c2114b7755075cf69e651d5e

Signed-off-by: Flopster101 <nahuelgomez329@gmail.com>