s5e8825-common: libaudio: Default Bluetooth voice sample rate to WB
A situation can occur where the bt_wbs parameter is passed too late and
the output remains set at NB (narrow band, 8 KHz), which will be percieved
as crunchy, aliased and often wrongly pitched audio in calls.

In my case, my headset was reporting WB support, as can be seen in logs:

$ adb logcat -b all | grep wbs
07-14 23:58:29.341   905   956 D audio_hw_primary: device-adev_set_parameters: enter with key(bt_headset_name=JBL Tune 520BT;bt_headset_nrec=off;bt_wbs=on)

Yet the HAL was not respecting it for this device.

A similar situation is described in this CLO commit:
https://git.codelinaro.org/clo/la/platform/vendor/qcom/opensource/arpal-lx/-/commit/bf1bd6cb1fc4ad19fd51eabe615aa224bcb65992

It also mentions "...most of BT hs supports WB mode...", which is why I
think this is a reasonable workaround for now.

Signed-off-by: Flopster101 <nahuelgomez329@gmail.com>