s5e8825-common: props: Set audio timecheck timeout to 10s
When trying to switch from Bluetooth to Speaker or Earpiece in a call, our
audio HAL takes a long time to perform the switch for some reason.

This can be seen in the logs:

07-17 23:32:29.109   900  2188 I audio_hw_primary: primary_out-adev_set_route-1: skip re-route as same device(bt-sco-headset) and same usage(voip_call)
07-17 23:32:29.109   900  2188 I audio_hw_primary: primary_out-adev_set_route-2: skip re-route as same device(bt-sco-headset-in) and same usage(voip_call)
07-17 23:32:29.109   900  2188 E audio_hw_primary: primary_in-in_set_parameters: This parameter cannot accept as Stream is working
07-17 23:32:29.111   900  2188 D audio_hw_primary: primary_in-in_update_sink_metadata: called, but not implemented yet
07-17 23:32:29.111   900  2188 D audio_hw_primary: primary_in-in_update_sink_metadata: This stream has 1 tracks

It occurs primarily in VoIP calls (e.g. Discord), but has also happened in
regular calls occasionally.

The time it takes to switch is about 7 seconds, and by default the TimeCheck
watchdog will kill it before that. Bumping to 10 seconds gives it enough time
to switch successfully.

NOTE that this is a workaround, but since I couldn't find a solution, it will
do for now.

Signed-off-by: Flopster101 <nahuelgomez329@gmail.com>